feat: Restricting concurrent editing

+++ refiner/app/api/v1/configurations.py
@@ -31,6 +31,8 @@ from app.db.configurations.activations.db import (
     activate_configuration_db,
     deactivate_configuration_db,
 )
+from app.db.users.db import get_user_by_id_db
+from app.services.configuration_locks import ConfigurationLock
 from app.services.configurations import (
     get_canonical_url_to_highest_inactive_version_map,
 )
@@ -269,6 +271,19 @@ async def create_configuration(
             detail="Unable to create configuration",
         )
 
+    # acquire lock for new configuration
+    success = await ConfigurationLock.acquire_lock(
+        configuration_id=config.id,
+        user_id=user.id,
+        jurisdiction_id=jd,
+        db=db,
+    )
+    if not success:
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail="Unable to acquire lock for new configuration",
+        )
+
     return CreateConfigurationResponse(id=config.id, name=config.name)
 
 
@@ -286,9 +301,20 @@ class IncludedCondition:
 
 
 @dataclass(frozen=True)
+class LockedByUser:
+    """
+    Model for the property lockedBy for GetConfigurationResponse.
+    """
+
+    id: str
+    name: str
+    email: str
+
+
+@dataclass(frozen=True)
 class GetConfigurationResponse:
     """
-    Information about a specific configuration to return to the client.
+    Model for a configration response.
     """
 
     id: UUID
@@ -307,6 +333,7 @@ class GetConfigurationResponse:
     version: int
     active_version: int | None
     latest_version: int
+    lockedBy: LockedByUser | None
 
 
 @dataclass(frozen=True)
@@ -421,6 +448,26 @@ async def get_configuration(
     active_version = active_config.version if active_config is not None else None
     latest_version = latest_config.version if latest_config is not None else 0
 
+    # get current lock
+    lock = await ConfigurationLock.get_lock(str(configuration_id), db)
+    if not lock or lock.expires_at <= datetime.utcnow():
+        await ConfigurationLock.acquire_lock(
+            configuration_id=configuration_id,
+            user_id=user.id,
+            jurisdiction_id=jd,
+            db=db,
+        )
+        lock = await ConfigurationLock.get_lock(str(configuration_id), db)
+    locked_by = None
+    if lock and lock.expires_at > datetime.utcnow():
+        try:
+            user_obj = await get_user_by_id_db(lock.user_id, db)
+            locked_by = LockedByUser(
+                id=str(user_obj.id), name=user_obj.username, email=user_obj.email
+            )
+        except Exception:
+            locked_by = None
+
     return GetConfigurationResponse(
         id=config.id,
         draft_id=draft_id,
@@ -438,6 +485,7 @@ async def get_configuration(
         version=config.version,
         active_version=active_version,
         latest_version=latest_version,
+        lockedBy=locked_by,
     )
 
 
@@ -593,6 +641,16 @@ async def associate_condition_codeset_with_configuration(
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Configuration not found."
         )
+    lock = await ConfigurationLock.get_lock(str(configuration_id), db)
+    if (
+        lock
+        and str(lock.user_id) != str(user.id)
+        and lock.expires_at > datetime.utcnow()
+    ):
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=f"{user.username}/{user.email} currently has this configuration open.",
+        )
 
     condition = await get_condition_by_id_db(id=body.condition_id, db=db)
 
@@ -661,6 +719,16 @@ async def remove_condition_codeset_from_configuration(
     if not config:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Configuration not found."
+        )
+    lock = await ConfigurationLock.get_lock(str(configuration_id), db)
+    if (
+        lock
+        and str(lock.user_id) != str(user.id)
+        and lock.expires_at > datetime.utcnow()
+    ):
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=f"{user.username}/{user.email} currently has this configuration open.",
         )
 
     condition = await get_condition_by_id_db(id=condition_id, db=db)
@@ -793,11 +861,27 @@ async def add_custom_code(
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Configuration not found."
         )
+    lock = await ConfigurationLock.get_lock(str(configuration_id), db)
+    if (
+        lock
+        and str(lock.user_id) != str(user.id)
+        and lock.expires_at > datetime.utcnow()
+    ):
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=f"{user.username}/{user.email} currently has this configuration open.",
+        )
 
     # Create a custom code object
+    allowed_systems = ["LOINC", "SNOMED", "ICD-10", "RxNorm"]
+    if sanitized_system_name not in allowed_systems:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail=f"System must be one of {allowed_systems}. Got: {sanitized_system_name}",
+        )
     custom_code = DbConfigurationCustomCode(
         code=body.code.strip(),
-        system=sanitized_system_name,
+        system=_get_literal_system(sanitized_system_name),
         name=body.name,
     )
 
@@ -879,6 +963,16 @@ async def delete_custom_code(
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Configuration not found."
         )
+    lock = await ConfigurationLock.get_lock(str(configuration_id), db)
+    if (
+        lock
+        and str(lock.user_id) != str(user.id)
+        and lock.expires_at > datetime.utcnow()
+    ):
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=f"{user.username}/{user.email} currently has this configuration open.",
+        )
 
     updated_config = await delete_custom_code_from_configuration_db(
         config=config, system=system, code=code, user_id=user.id, db=db
@@ -952,12 +1046,21 @@ def _get_modified_custom_codes(
 
     # create a new code using the changes provided by the user.
     # use the old values as fallbacks.
+    allowed_systems = ["LOINC", "SNOMED", "ICD-10", "RxNorm"]
+    new_system_value = (
+        _get_sanitized_system_name(updateInput.new_system)
+        if updateInput.new_system
+        else existing_code.system
+    )
+    if new_system_value not in allowed_systems:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail=f"System must be one of {allowed_systems}. Got: {new_system_value}",
+        )
     updated_code = DbConfigurationCustomCode(
         code=updateInput.new_code or existing_code.code,
         name=updateInput.new_name or existing_code.name,
-        system=_get_sanitized_system_name(updateInput.new_system)
-        if updateInput.new_system
-        else existing_code.system,
+        system=_get_literal_system(new_system_value),
     )
 
     # check for duplicates
@@ -1038,6 +1141,16 @@ async def edit_custom_code(
     if not config:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Configuration not found."
+        )
+    lock = await ConfigurationLock.get_lock(str(configuration_id), db)
+    if (
+        lock
+        and str(lock.user_id) != str(user.id)
+        and lock.expires_at > datetime.utcnow()
+    ):
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=f"{user.username}/{user.email} currently has this configuration open.",
         )
 
     custom_codes = _get_modified_custom_codes(
@@ -1426,6 +1539,17 @@ async def update_section_processing(
             status_code=status.HTTP_404_NOT_FOUND, detail="Configuration not found."
         )
 
+    lock = await ConfigurationLock.get_lock(str(configuration_id), db)
+    if (
+        lock
+        and str(lock.user_id) != str(user.id)
+        and lock.expires_at > datetime.utcnow()
+    ):
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=f"{user.username}/{user.email} currently has this configuration open.",
+        )
+
     # convert payload to DB-friendly format (SectionUpdate dataclasses)
     section_updates = [
         SectionUpdate(code=s.code, action=s.action) for s in payload.sections
@@ -1585,3 +1709,18 @@ async def deactivate_configuration(
         condition_canonical_url=deactivated_config.condition_canonical_url,
         status="inactive",
     )
+
+
+def _get_literal_system(system: str) -> Literal["LOINC", "SNOMED", "ICD-10", "RxNorm"]:
+    """
+    Helper to ensure Literal type for custom code system.
+    """
+    if system == "LOINC":
+        return "LOINC"
+    if system == "SNOMED":
+        return "SNOMED"
+    if system == "ICD-10":
+        return "ICD-10"
+    if system == "RxNorm":
+        return "RxNorm"
+    raise ValueError(f"Invalid system: {system}")
